Transakcija logic changed

+++ mybank/mybank-api/src/main/java/com/project/dtos/racun/RacunDepositDto.java
@@ -15,5 +15,5 @@ public class RacunDepositDto {
     @NotNull
     @Min(message = "iznos mora biti veci od 10", value = 10)
     @Max(message = "iznos ne moze biti veci od 100000", value = 100000)
-    private int iznos;
+    private double iznos;
 }
