+++ mybank/mybank-api/src/main/java/com/project/dtos/racun/RacunWithdrawDto.java
@@ -15,5 +15,5 @@ public class RacunWithdrawDto {
     @NotNull
     @Min(message = "Iznos ne moze biti manji od 0", value = 0)
     @Max(message = "Iznos ne moze biti veci od 100000", value = 100000)
-    private int iznos;
+    private double iznos;
 }
